reorganization * split projects into driver (main.cpp) and core logic library (including runtime) * adapted paths to fetch runtime files from * updated documentation
