+++ src/lib/IndentingStreamWrapper.h
