+++ src/lib/NotImplementedException.h
