+++ src/lib/Variant.hpp
